+++ diabeticat-ui/src/pages/PetPage.vue
@@ -1,0 +1,15 @@
+<template>
+    <PetCard></PetCard>
+    <MeasurementData></MeasurementData>
+</template>
+  
+  <script setup>
+    import PetCard from '../components/PetCard.vue'
+    import MeasurementData from '../components/MeasurementData.vue'
+
+  </script>
+  
+  <style scoped>
+  
+  </style>
+
